refactor: extract browser evidence runtime
